split event processing
